requirements.txt
# OCR and PDF processing
ocrmypdf>=3.3.0
pypdf>=3.0.0

# LangChain ecosystem
langchain>=0.1.0
langchain-core>=0.1.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1
langchain-ollama>=0.1.0

# Vector database
chromadb>=0.4.0

# Ollama integration
ollama>=0.1.0

# Environment management
python-dotenv>=1.0.0

# Streamlit web interface
streamlit>=1.28.0

# Additional dependencies for PDF processing
Pillow>=9.0.0
tesseract>=0.1.3

# Optional: For better PDF handling
pdfplumber>=0.9.0
